SDK REL.MCUSDK.26.01.00.07 Release Package
List of changes from previous build

## mcu_plus_sdk
Previous: 0736e0c30b05 -> Current: cd6e960cbb9f

cd6e960cbb9 am26xx: docs: update release note
3e9baeadd63 am261x: bootloader: fix sbl psram dac
ae715e1c7e0 am26xx: docs: update custom board docs in enet migration guide
370cf13bc7e am26xx: docs: update Networking migration for syscfg 1.28.0
b2589f1cbfd am26xx: docs: update RN Features section
1075f597940 am26xx: docs: update RN
e8a2f4dd1cf am26xx: docs: Update datasheet
c65aacc024e enet: am263x/am263px/am261x: Update Custom-PHY migration guide
964e0e43960 enet: am263x/am263px/am261x: Update ethernet performance page
50a9ef9af72 am263x: sdl: enable cacheP after pbist test
48f94b2a459 am26xx: sdl: fix klocwork warning
2808c56b785 am273x: Docs: Fix docs error
1e40189b482 am263x/am263px: kernel: Klocwork fix on kernel
096c9b242cd am26xx: pbist: remove unused api

## security_common
Previous: 5ab89e3e1381 -> Current: 8695773f6fba

8695773 f29h85x: secure_ipc_notify: misra fixes
29d1a0a f29x: rng/pke: Add design ID
d09ea58 f29x: sipc: Add design ID
065c936 f29x: hsmclient: Add design ID

## icss_emac_lld
Previous: af833cfa49e6 -> Current: 9f2da368d4d7

fba7644 allsoc: icss emac: add DebugP_assert check for invalid init params
753be72 allsoc: icss emac/timesync: standardize spacing around comparison and assignment operators
795c5d1 allsoc: icss emac/timesync: standardize keyword and operator spacing
8b72329 allsoc: icss_timesync: enforce lower bound on ptpGeneralSize in memcpy branch
99c29a7 allsoc: icss_emac: address Klockwork array index tracking vulnerabilities

+++ source/security/security_common/drivers/secure_ipc_notify/sipc_notify_mailbox.h
@@ -148,7 +148,7 @@ static inline int32_t SIPC_mailboxWrite(uint32_t mailboxBaseAddr, uint32_t wrInt
             #endif
 
             /* Trigger interrupt to other core */
-            *addr = (1U << (wrIntrBitPos));
+            *addr = ((uint32_t)1 << (wrIntrBitPos));
 
             status = SystemP_SUCCESS;
         }
